requirements.txt
attrs==22.1.0
blinker==1.5
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
contourpy==1.0.6
cryptography==38.0.1
cycler==0.11.0
dnspython==2.2.1
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Mail==0.9.1
fonttools==4.38.0
idna==3.4
importlib-metadata==5.0.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
motor==3.0.0
numpy==1.23.3
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
pluggy==1.0.0
py==1.11.0
pycparser==2.21
PyJWT==2.5.0
pymongo==4.2.0
pyparsing==3.0.9
pytest==7.1.3
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2022.6
requests==2.28.1
seaborn==0.12.1
six==1.16.0
tomli==2.0.1
urllib3==1.26.12
Werkzeug==2.2.2
wincertstore==0.2
zipp==3.8.1
